refactor: some service layers were converted to a more moduler file structure
