+++ src/hooks/use-products.ts
@@ -100,58 +100,3 @@ export const useDeleteProduct = () => {
     },
   });
 };
-
-// Product Question Answers hooks
-export const useProductQuestionAnswers = (productId: string) => {
-  return useQuery({
-    queryKey: ["product-question-answers", productId],
-    queryFn: () => productsService.getProductQuestionAnswers(productId),
-    enabled: !!productId,
-  });
-};
-
-export const useCreateProductQuestionAnswers = () => {
-  const queryClient = useQueryClient();
-
-  return useMutation({
-    mutationFn: ({
-      productId,
-      questionAnswers,
-    }: {
-      productId: string;
-      questionAnswers: Array<{ question_id: string; answer: boolean }>;
-    }) =>
-      productsService.createProductQuestionAnswers(productId, questionAnswers),
-    onSuccess: (_, variables) => {
-      queryClient.invalidateQueries({
-        queryKey: ["product-question-answers", variables.productId],
-        refetchType: "all",
-      });
-    },
-  });
-};
-
-export function useDeleteProductQuestionAnswersByIds() {
-  const queryClient = useQueryClient();
-
-  return useMutation({
-    mutationFn: async ({
-      productId,
-      questionIds,
-    }: {
-      productId: string;
-      questionIds: string[];
-    }) => {
-      await productsService.deleteProductQuestionAnswersByIds(
-        productId,
-        questionIds
-      );
-    },
-    onSuccess: (_, variables) => {
-      queryClient.invalidateQueries({
-        queryKey: ["product-question-answers", variables.productId],
-        refetchType: "all",
-      });
-    },
-  });
-}
